requirements.txt
streamlit
beautifulsoup4
requests
jinja2
selenium
webdriver-manager
python-docx
pypandoc
